% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_project.R
\name{setup_with_git}
\alias{setup_with_git}
\title{Setup Git to the project.}
\usage{
setup_with_git()
}
\value{
Adds Git and `.gitignore` file to the project.
}
\description{
Takes a lot of inspiration from the usethis `use_git()` function, except
it only adds Git and nothing more (doesn't commit, doesn't restart RStudio
automatically). Must run this function inside the project you created from
[setup_project()]
}
\seealso{
[setup_project()] for starting the project.
}
